requirements.txt
Django==6.0.3
asgiref>=3.9.1
sqlparse>=0.5.0
tzdata>=2024.1
djangorestframework>=3.15.0